requirements.txt
agora-python-sdk==3.1.2.10
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ecdsa==0.17.0
enum34==1.1.10
gunicorn==20.1.0
idna==3.3
opentok==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
PyJWT==2.4.0
python-jose==3.3.0
pytz==2022.1
requests==2.28.0
rsa==4.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
